[fix] move make command to sub directory
