+++ src/Commands/Make/MailMakeCommand.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace Nwidart\Modules\Commands;
+namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Support\Str;
+use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 
-final class ChannelMakeCommand extends GeneratorCommand
+class MailMakeCommand extends GeneratorCommand
 {
     use ModuleCommandTrait;
 
@@ -17,21 +18,34 @@ final class ChannelMakeCommand extends GeneratorCommand
      *
      * @var string
      */
-    protected $name = 'module:make-channel';
-
-    protected $argumentName = 'name';
+    protected $name = 'module:make-mail';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Create a new channel class for the specified module.';
+    protected $description = 'Create a new email class for the specified module';
+
+    protected $argumentName = 'name';
 
     public function getDefaultNamespace(): string
     {
-        return config('modules.paths.generator.channels.namespace')
-            ?? ltrim(config('modules.paths.generator.channels.path', 'Broadcasting'), config('modules.paths.app_folder', ''));
+        return config('modules.paths.generator.emails.namespace')
+            ?? ltrim(config('modules.paths.generator.emails.path', 'Emails'), config('modules.paths.app_folder', ''));
+    }
+
+    /**
+     * Get the console command arguments.
+     *
+     * @return array
+     */
+    protected function getArguments()
+    {
+        return [
+            ['name', InputArgument::REQUIRED, 'The name of the mailable.'],
+            ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
+        ];
     }
 
     /**
@@ -43,7 +57,7 @@ final class ChannelMakeCommand extends GeneratorCommand
     {
         $module = $this->laravel['modules']->findOrFail($this->getModuleName());
 
-        return (new Stub('/channel.stub', [
+        return (new Stub('/mail.stub', [
             'NAMESPACE' => $this->getClassNamespace($module),
             'CLASS'     => $this->getClass(),
         ]))->render();
@@ -58,9 +72,9 @@ final class ChannelMakeCommand extends GeneratorCommand
     {
         $path = $this->laravel['modules']->getModulePath($this->getModuleName());
 
-        $channelPath = GenerateConfigReader::read('channels');
+        $mailPath = GenerateConfigReader::read('emails');
 
-        return $path . $channelPath->getPath() . '/' . $this->getFileName() . '.php';
+        return $path . $mailPath->getPath() . '/' . $this->getFileName() . '.php';
     }
 
     /**
@@ -69,18 +83,5 @@ final class ChannelMakeCommand extends GeneratorCommand
     private function getFileName()
     {
         return Str::studly($this->argument('name'));
-    }
-
-    /**
-     * Get the console command arguments.
-     *
-     * @return array
-     */
-    protected function getArguments()
-    {
-        return [
-            ['name', InputArgument::REQUIRED, 'The name of the channel class.'],
-            ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
-        ];
     }
 }
