+++ src/Commands/Make/RequestMakeCommand.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace Nwidart\Modules\Commands;
+namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Support\Str;
+use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 
-class PolicyMakeCommand extends GeneratorCommand
+class RequestMakeCommand extends GeneratorCommand
 {
     use ModuleCommandTrait;
 
@@ -24,19 +25,19 @@ class PolicyMakeCommand extends GeneratorCommand
      *
      * @var string
      */
-    protected $name = 'module:make-policy';
+    protected $name = 'module:make-request';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Create a new policy class for the specified module.';
+    protected $description = 'Create a new form request class for the specified module.';
 
     public function getDefaultNamespace(): string
     {
-        return config('modules.paths.generator.policies.namespace')
-            ?? ltrim(config('modules.paths.generator.policies.path', 'Policies'), config('modules.paths.app_folder', ''));
+        return config('modules.paths.generator.request.namespace')
+            ?? ltrim(config('modules.paths.generator.request.path', 'Http/Requests'), config('modules.paths.app_folder', ''));
     }
 
     /**
@@ -47,7 +48,7 @@ class PolicyMakeCommand extends GeneratorCommand
     protected function getArguments()
     {
         return [
-            ['name', InputArgument::REQUIRED, 'The name of the policy class.'],
+            ['name', InputArgument::REQUIRED, 'The name of the form request class.'],
             ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
         ];
     }
@@ -59,7 +60,7 @@ class PolicyMakeCommand extends GeneratorCommand
     {
         $module = $this->laravel['modules']->findOrFail($this->getModuleName());
 
-        return (new Stub('/policy.plain.stub', [
+        return (new Stub('/request.stub', [
             'NAMESPACE' => $this->getClassNamespace($module),
             'CLASS'     => $this->getClass(),
         ]))->render();
@@ -72,9 +73,9 @@ class PolicyMakeCommand extends GeneratorCommand
     {
         $path = $this->laravel['modules']->getModulePath($this->getModuleName());
 
-        $policyPath = GenerateConfigReader::read('policies');
+        $requestPath = GenerateConfigReader::read('request');
 
-        return $path . $policyPath->getPath() . '/' . $this->getFileName() . '.php';
+        return $path . $requestPath->getPath() . '/' . $this->getFileName() . '.php';
     }
 
     /**
