+++ src/Commands/Make/ResourceMakeCommand.php
@@ -1,31 +1,27 @@
 <?php
 
-namespace Nwidart\Modules\Commands;
+namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Support\Str;
+use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
-class TestMakeCommand extends GeneratorCommand
+class ResourceMakeCommand extends GeneratorCommand
 {
     use ModuleCommandTrait;
 
     protected $argumentName = 'name';
-    protected $name = 'module:make-test';
-    protected $description = 'Create a new test class for the specified module.';
+    protected $name = 'module:make-resource';
+    protected $description = 'Create a new resource class for the specified module.';
 
     public function getDefaultNamespace(): string
     {
-        if ($this->option('feature')) {
-            return config('modules.paths.generator.test-feature.namespace')
-                ?? config('modules.paths.generator.test-feature.path', 'Tests/Feature');
-        }
-
-        return config('modules.paths.generator.test-unit.namespace')
-            ?? config('modules.paths.generator.test-unit.path', 'Tests/Unit');
+        return config('modules.paths.generator.resource.namespace')
+            ?? ltrim(config('modules.paths.generator.resource.path', 'Transformers'), config('modules.paths.app_folder', ''));
     }
 
     /**
@@ -36,20 +32,15 @@ class TestMakeCommand extends GeneratorCommand
     protected function getArguments()
     {
         return [
-            ['name', InputArgument::REQUIRED, 'The name of the form request class.'],
+            ['name', InputArgument::REQUIRED, 'The name of the resource class.'],
             ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
         ];
     }
 
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
     protected function getOptions()
     {
         return [
-            ['feature', null, InputOption::VALUE_NONE, 'Create a feature test.'],
+            ['collection', 'c', InputOption::VALUE_NONE, 'Create a resource collection.'],
         ];
     }
 
@@ -59,13 +50,8 @@ class TestMakeCommand extends GeneratorCommand
     protected function getTemplateContents()
     {
         $module = $this->laravel['modules']->findOrFail($this->getModuleName());
-        $stub = '/unit-test.stub';
 
-        if ($this->option('feature')) {
-            $stub = '/feature-test.stub';
-        }
-
-        return (new Stub($stub, [
+        return (new Stub($this->getStubName(), [
             'NAMESPACE' => $this->getClassNamespace($module),
             'CLASS'     => $this->getClass(),
         ]))->render();
@@ -78,13 +64,9 @@ class TestMakeCommand extends GeneratorCommand
     {
         $path = $this->laravel['modules']->getModulePath($this->getModuleName());
 
-        if ($this->option('feature')) {
-            $testPath = GenerateConfigReader::read('test-feature');
-        } else {
-            $testPath = GenerateConfigReader::read('test');
-        }
+        $resourcePath = GenerateConfigReader::read('resource');
 
-        return $path . $testPath->getPath() . '/' . $this->getFileName() . '.php';
+        return $path . $resourcePath->getPath() . '/' . $this->getFileName() . '.php';
     }
 
     /**
@@ -93,5 +75,28 @@ class TestMakeCommand extends GeneratorCommand
     private function getFileName()
     {
         return Str::studly($this->argument('name'));
+    }
+
+    /**
+     * Determine if the command is generating a resource collection.
+     *
+     * @return bool
+     */
+    protected function collection(): bool
+    {
+        return $this->option('collection') ||
+            Str::endsWith($this->argument('name'), 'Collection');
+    }
+
+    /**
+     * @return string
+     */
+    protected function getStubName(): string
+    {
+        if ($this->collection()) {
+            return '/resource-collection.stub';
+        }
+
+        return '/resource.stub';
     }
 }
