+++ src/Commands/Make/RuleMakeCommand.php
@@ -1,38 +1,44 @@
 <?php
 
-namespace Nwidart\Modules\Commands;
+namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Support\Str;
+use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
-class JobMakeCommand extends GeneratorCommand
+class RuleMakeCommand extends GeneratorCommand
 {
     use ModuleCommandTrait;
+
+    /**
+     * The name of argument name.
+     *
+     * @var string
+     */
+    protected $argumentName = 'name';
 
     /**
      * The console command name.
      *
      * @var string
      */
-    protected $name = 'module:make-job';
+    protected $name = 'module:make-rule';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Create a new job class for the specified module';
-
-    protected $argumentName = 'name';
+    protected $description = 'Create a new validation rule for the specified module.';
 
     public function getDefaultNamespace(): string
     {
-        return config('modules.paths.generator.jobs.namespace')
-            ?? ltrim(config('modules.paths.generator.jobs.path', 'Jobs'), config('modules.paths.app_folder', ''));
+        return config('modules.paths.generator.rules.namespace')
+            ?? ltrim(config('modules.paths.generator.rules.path', 'Rules'), config('modules.paths.app_folder', ''));
     }
 
     /**
@@ -43,7 +49,7 @@ class JobMakeCommand extends GeneratorCommand
     protected function getArguments()
     {
         return [
-            ['name', InputArgument::REQUIRED, 'The name of the job.'],
+            ['name', InputArgument::REQUIRED, 'The name of the rule class.'],
             ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
         ];
     }
@@ -56,37 +62,37 @@ class JobMakeCommand extends GeneratorCommand
     protected function getOptions()
     {
         return [
-            ['sync', null, InputOption::VALUE_NONE, 'Indicates that job should be synchronous.'],
+            ['implicit', 'i', InputOption::VALUE_NONE, 'Generate an implicit rule'],
         ];
     }
 
     /**
-     * Get template contents.
-     *
-     * @return string
+     * @return mixed
      */
     protected function getTemplateContents()
     {
         $module = $this->laravel['modules']->findOrFail($this->getModuleName());
 
-        return (new Stub($this->getStubName(), [
+        $stub = $this->option('implicit')
+            ? '/rule.implicit.stub'
+            : '/rule.stub';
+
+        return (new Stub($stub, [
             'NAMESPACE' => $this->getClassNamespace($module),
-            'CLASS'     => $this->getClass(),
+            'CLASS'     => $this->getFileName(),
         ]))->render();
     }
 
     /**
-     * Get the destination file path.
-     *
-     * @return string
+     * @return mixed
      */
     protected function getDestinationFilePath()
     {
         $path = $this->laravel['modules']->getModulePath($this->getModuleName());
 
-        $jobPath = GenerateConfigReader::read('jobs');
+        $rulePath = GenerateConfigReader::read('rules');
 
-        return $path . $jobPath->getPath() . '/' . $this->getFileName() . '.php';
+        return $path . $rulePath->getPath() . '/' . $this->getFileName() . '.php';
     }
 
     /**
@@ -95,17 +101,5 @@ class JobMakeCommand extends GeneratorCommand
     private function getFileName()
     {
         return Str::studly($this->argument('name'));
-    }
-
-    /**
-     * @return string
-     */
-    protected function getStubName(): string
-    {
-        if ($this->option('sync')) {
-            return '/job.stub';
-        }
-
-        return '/job-queued.stub';
     }
 }
